+++ Project/src/main/java/com/koreait/project/dto/Board_knowhow_tagDto.java
@@ -1,16 +1,17 @@
-package com.koreait.project.Dto;
+package com.koreait.project.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+@Data
 @AllArgsConstructor
 @NoArgsConstructor
-@Data
-public class ExerciseDto {
+public class Board_knowhow_tagDto {
 
-	private int exercise_no;
-	private String exercise_name;
+	private int knowhow_tag_no;
+	private int knowhow_no;
+	private int tag_no;
 	
 	/*
 	 * 0 : 족구
@@ -22,6 +23,9 @@ public class ExerciseDto {
 	 * 6 : 댄스
 	 * 7 : 요가
 	 * 8 : 명상
+	 * 9 : 다이어트정보
+	 * 10 : 건강정보
+	 * 11 : 식단정보
 	 */
 	
 }
