+++ Project/src/main/java/com/koreait/project/dto/ExerciseDto.java
@@ -1,17 +1,16 @@
-package com.koreait.project.Dto;
+package com.koreait.project.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-@Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class Board_knowhow_tagDto {
+@Data
+public class ExerciseDto {
 
-	private int knowhow_tag_no;
-	private int knowhow_no;
-	private int tag_no;
+	private int exercise_no;
+	private String exercise_name;
 	
 	/*
 	 * 0 : 족구
@@ -23,9 +22,6 @@ public class Board_knowhow_tagDto {
 	 * 6 : 댄스
 	 * 7 : 요가
 	 * 8 : 명상
-	 * 9 : 다이어트정보
-	 * 10 : 건강정보
-	 * 11 : 식단정보
 	 */
 	
 }
