+++ Project/src/main/java/com/koreait/project/dto/Trainer_qnaDto.java
@@ -1,4 +1,4 @@
-package com.koreait.project.Dto;
+package com.koreait.project.dto;
 
 import java.sql.Date;
